Worked on DataGen, fixed MOST of the item models

+++ src/main/java/com/code/tama/tts/server/blocks/HartnellDoor.java
@@ -4,6 +4,8 @@ package com.code.tama.tts.server.blocks;
 import com.code.tama.tts.server.registries.TTSBlocks;
 import com.code.tama.tts.server.tileentities.HartnellDoorTile;
 import com.code.tama.tts.server.tileentities.HartnellDoorTilePlaceholder;
+import java.util.function.Supplier;
+import javax.annotation.Nullable;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundEvents;
@@ -26,9 +28,6 @@ import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.NotNull;
 
-import javax.annotation.Nullable;
-import java.util.function.Supplier;
-
 @SuppressWarnings("deprecation")
 public class HartnellDoor extends Block implements EntityBlock {
     private boolean IsOpen = false;
@@ -49,8 +48,7 @@ public class HartnellDoor extends Block implements EntityBlock {
 
     @Override
     public void destroy(LevelAccessor levelAccessor, @NotNull BlockPos blockPos, @NotNull BlockState state) {
-        if(!levelAccessor.isClientSide())
-            this.destroyMultiblockStructure((ServerLevel) levelAccessor, blockPos);
+        if (!levelAccessor.isClientSide()) this.destroyMultiblockStructure((ServerLevel) levelAccessor, blockPos);
         super.destroy(levelAccessor, blockPos, state);
     }
 
@@ -74,7 +72,10 @@ public class HartnellDoor extends Block implements EntityBlock {
 
     @Override
     public @NotNull VoxelShape getCollisionShape(
-            @NotNull BlockState state, @NotNull BlockGetter blockGetter, @NotNull BlockPos blockPos, @NotNull CollisionContext collisionContext) {
+            @NotNull BlockState state,
+            @NotNull BlockGetter blockGetter,
+            @NotNull BlockPos blockPos,
+            @NotNull CollisionContext collisionContext) {
         if (this.IsOpen) return Shapes.empty();
         return super.getCollisionShape(state, blockGetter, blockPos, collisionContext);
     }
